feat: admin을 위한 letters store 생성
